decodyService/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.4
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
coverage==7.6.4
dill==0.3.8
distro==1.9.0
Flask==3.0.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mccabe==0.7.0
openai==1.55.1
platformdirs==4.3.6
pydantic==2.10.2
pydantic_core==2.27.1
pylint==3.3.1
PyMySQL==1.1.1
python-dotenv==1.0.1
referencing==0.35.1
rpds-py==0.20.0
sniffio==1.3.1
tomlkit==0.13.2
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.0.4
pyyaml==6.0.2
